geodemo/requirements.txt
Django==1.11.13
djangorestframework==3.8.2
djangorestframework==0.13
Markdown==2.6.2
django-filter==0.11.0
psycopg2==2.6.2
django-cors-headers==1.1.0